edit fixed, drag/drop applied, dashboard - msg/tasks

+++ angular-office-board/client/src/Client/src/app/tasks/task-baord/task-baord.component.ts
@@ -3,6 +3,8 @@ import { Router } from '@angular/router';
 import { Task, User } from '../../shared/interfaces';
 import { TaskService } from '../task.service';
 import { AuthService } from 'src/app/auth/auth.service';
+import { CdkDragDrop, moveItemInArray, transferArrayItem } from '@angular/cdk/drag-drop';
+
 
 @Component({
   selector: 'app-task-baord',
@@ -10,15 +12,15 @@ import { AuthService } from 'src/app/auth/auth.service';
   styleUrls: ['./task-baord.component.scss'],
 })
 export class TaskBaordComponent implements OnInit {
-  toDoTasks: Array<Task>;
-  doingTasks: Array<Task>;
-  doneTasks: Array<Task>;
+  toDoTasks: Array<Task> = [];
+  doingTasks: Array<Task> = [];
+  doneTasks: Array<Task> = [];
   currentUser!: User;
 
   constructor(private taskService: TaskService, private router: Router, private authService: AuthService) {
-    this.toDoTasks = new Array<Task>();
-    this.doingTasks = new Array<Task>();
-    this.doneTasks = new Array<Task>();
+    // this.toDoTasks = new Array<Task>();
+    // this.doingTasks = new Array<Task>();
+    // this.doneTasks = new Array<Task>();
 
     this.authService.getCurrentUser().subscribe((res) => {
       this.currentUser = res;
@@ -32,26 +34,34 @@ export class TaskBaordComponent implements OnInit {
 
   fetchTasks() {
     this.taskService.getAllTasks().subscribe((res) => {
-      this.toDoTasks = res.filter((x) => x.status == 1);
-      this.doingTasks = res.filter((x) => x.status == 2);
-      this.doneTasks = res.filter((x) => x.status == 3);
+      
+      if (res !== undefined) {
+        return (
+
+          this.toDoTasks = res.filter((x) => x.status == 1),
+          this.doingTasks = res.filter((x) => x.status == 2),
+          this.doneTasks = res.filter((x) => x.status == 3)
+          )
+        }
+        return null
+
     });    
   };
 
   
-  deleteTask(id: number) {
-    this.taskService.deleteTask(id).subscribe(res => {
+  deleteTask(id: string) {
+    this.taskService.deleteTask(id).subscribe(res => {      
       this.fetchTasks();
     });
   }
 
-  confirmDelete(name: string, id: number) {
-    if (confirm(`${name} - Delete task?`)) {
+  confirmDelete(name: string, id: string | undefined) {
+    if (confirm(`${name} - Delete task?`) && id) {
       this.deleteTask(id);
     }
   }
 
-  changeStatus(id: number, status: number) {
+  changeStatus(id: string | undefined, status: number) {
     const data = {
       'status': status
     }
@@ -60,5 +70,14 @@ export class TaskBaordComponent implements OnInit {
     });
   }
 
+  drop(event: CdkDragDrop<Task[]>, status: number) {
+    if (event.previousContainer !== event.container) {
+      transferArrayItem(event.previousContainer.data, event.container.data,event.previousIndex, event.currentIndex);
+      let taskId: string | undefined = event.item.getRootElement().querySelector('#taskId')?.innerHTML
+        this.changeStatus(taskId, status);
+    } else {
+      moveItemInArray(this.toDoTasks, event.previousIndex, event.currentIndex);
+    }
+  }
 
 }
